fix: small bugs
Mostly functionality that is still missing from my commits

+++ tests/classes/conftest.py
@@ -65,7 +65,7 @@ def manual_slab_3_atoms():
     slab = Slab()
     slab.ucell = np.diag([3., 4., 5.])
     positions = (-0.25, 0, 0), (0, 0, 0), (0.25, 0, 0)
-    slab.atlist.extend(Atom('C', np.array(pos), i+1, slab)
+    slab.atlist.extend(Atom('C', np.array(pos, dtype=float), i+1, slab)
                        for i, pos in enumerate(positions))
     slab.full_update(Rparams())
     return slab
